fix: Autofix and enable lints
* Fix the lints, which auto fixable with golangci-lint run --fix
* Enable security (gosec) lint
* Enable performance (prealloc) lint
* Ensure license headers

+++ server/metrics/metrics.go
@@ -67,7 +67,7 @@ func getTimerSummaryObjectives() map[float64]float64 {
 	res := make(map[float64]float64)
 	for confQuantile, objective := range getTigrisDefaultSummaryObjectives() {
 		for _, wantedQuantile := range config.DefaultConfig.Metrics.TimerQuantiles {
-			if confQuantile == float64(wantedQuantile) {
+			if confQuantile == wantedQuantile {
 				res[confQuantile] = objective
 			}
 		}
